create entry folder, move register and login
